fix: Handle case when user is not found in DeleteUserUsecase
This commit adds error handling in the DeleteUserUsecase to throw a NotFoundException when the user with the specified ID is not found. This prevents potential errors and provides better feedback to the client.

+++ src/@core/application/usecases/user/delete-user.usecase.ts
@@ -1,10 +1,12 @@
 import { IUserRepository } from '../../../domain/repositories/user.repository';
+import { NotFoundException } from '../../../domain/shared/errors/not-found.exception';
 
 export class DeleteUserUsecase {
   constructor(private readonly userRepository: IUserRepository) {}
 
   async execute(id: number): Promise<void> {
     const user = await this.userRepository.findById(id);
+    if (!user) throw new NotFoundException('USER_NOT_FOUND');
     await this.userRepository.delete(user);
     return;
   }
